DEV: organize project structure for modularize python project
