+++ src/quant/structure.py
